pyproject.toml
[project]
name = "agent-w-mem"
version = "0.1.0"
description = "Azure OpenAI Agent with Graphiti Temporal Knowledge Graph Memory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "graphiti-core>=0.1.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
    "pydantic>=2.0.0"
]
